mofification fonction distance

+++ point.c
@@ -44,7 +44,7 @@ float distanceEntreDeuxPoints(point p1, point p2)
   double x2 = getX(p2);
   double y1 = getY(p1);
   double y2 = getY(p2);
-  double tmp =  (x2 - x1) + (y2 - y1);
+  float tmp =  (x2 - x1) + (y2 - y1);
 	if (tmp<0)
 		tmp=-tmp;
   return tmp;
